packages/bu-tap-qualtrics/pyproject.toml
[tool.poetry]
name = "bu-tap-qualtrics"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "bu_tap_qualtrics"}]

[tool.poetry.dependencies]
python = "^3.10"
singer-python = "^5.13.0"
requests = "^2.28.2"
requests-oauthlib = "^1.3.1"
python-dotenv = "^0.21.1"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
mock = "^5.0.1"

[tool.poetry.scripts]
bu-tap-qualtrics = "bu_tap_qualtrics:main"
create-qualtrics-configs = "bu_tap_qualtrics.configs:create_qualtrics_configs"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
